#22 Move internal methods from ChannelLike to ChannelImpl interface.
